fix: move file controller to folder
